+++ backend/src/main/java/wooteco/prolog/login/ui/AuthMemberPrincipalArgumentResolver.java
@@ -1,14 +1,41 @@
 package wooteco.prolog.login.ui;
 
+import lombok.AllArgsConstructor;
 import org.springframework.core.MethodParameter;
+import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import wooteco.prolog.login.application.AuthorizationExtractor;
+import wooteco.prolog.login.application.JwtTokenProvider;
+import wooteco.prolog.login.application.MemberService;
+import wooteco.prolog.login.domain.AuthMemberPrincipal;
 import wooteco.prolog.login.domain.Member;
+import wooteco.prolog.login.excetpion.TokenNotValidException;
 
-public interface AuthMemberPrincipalArgumentResolver extends HandlerMethodArgumentResolver {
-    boolean supportsParameter(MethodParameter parameter);
+import javax.servlet.http.HttpServletRequest;
 
-    Member resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory);
+@AllArgsConstructor
+@Component
+public class AuthMemberPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
+
+    private final MemberService memberService;
+    private final JwtTokenProvider jwtTokenProvider;
+
+    @Override
+    public boolean supportsParameter(MethodParameter parameter) {
+        return parameter.hasParameterAnnotation(AuthMemberPrincipal.class);
+    }
+
+    @Override
+    public Member resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+        String credentials = AuthorizationExtractor.extract(webRequest.getNativeRequest(HttpServletRequest.class));
+        try {
+            Long id = Long.valueOf(jwtTokenProvider.extractSubject(credentials));
+            return memberService.findById(id);
+        } catch (NumberFormatException e) {
+            throw new TokenNotValidException();
+        }
+    }
 }
