+++ backend/src/test/resources/wooteco/prolog/acceptance/post.feature
@@ -30,7 +30,7 @@ Feature: 포스트 관련 기능
     When 1번째 포스트를 조회하면
     Then 1번째 포스트가 조회된다
 
-#  Scenario: 포스트 삭제하기
-#    * 포스트 여러개를 작성하고
-#    * 1번째 포스트를 삭제하면
-#    * 1번째 포스트가 삭제된다
+  Scenario: 포스트 삭제하기
+    * 포스트 여러개를 작성하고
+    * 1번째 포스트를 삭제하면
+    * 1번째 포스트가 삭제된다
